Handle graphs with weights as floats + fix Makefile

+++ dijkstra.h
@@ -1,3 +1,3 @@
 #include "graph.h"
 
-void dijkstra(graph_t* G, int s, int* prev, int* dist);
+void dijkstra(graph_t* G, int s, int* prev, double* dist);
