+++ graph.h
@@ -6,7 +6,7 @@ typedef struct states states_t;
 
 typedef struct adjacency_list_node {  // Represents edges between nodes
   int vertex;  // Other end of the edge
-  int weight;  // Weight of the edge
+  double weight;  // Weight of the edge
   struct adjacency_list_node *next;  // Pointer to next item in the adjacency linked list
 } adjacency_list_node_t;
 
@@ -24,7 +24,7 @@ typedef struct graph {
 graph_t *createGraph(int n);
 void freeGraph(graph_t *graph);
 int *getNodes(graph_t const *graph);
-void addEdge(graph_t *graph, int src, int dest, int weight);
+void addEdge(graph_t *graph, int src, int dest, double weight);
 void printGraph(graph_t const *graph);
 
 graph_t *reverseGraph(graph_t const *graph);
